LayoutTests/http/tests/security/sandboxed-iframe-geolocation-getCurrentPosition-expected.txt

CONSOLE MESSAGE: Feature policy 'Geolocation' check failed for iframe with origin 'null' and allow attribute ''.
Tests that navigator.geolocation.getCurrentPosition() returns error PERMISSION_DENIED when called from a document in a sandboxed iframe.



--------
Frame: '<!--frame1-->'
--------
PASS error.code is error.PERMISSION_DENIED.
PASS error.message is "Origin does not have permission to use Geolocation service".
